pyproject.toml
[project]
name = "finance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.6",
    "langchain-openai>=1.1.2",
    "mcp-use>=1.5.0",
    "openai>=2.11.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = true

[tool.uv.sources]
finance_agent = { path = "src/finance_agent" }
